Refactor Add Candy menu to new file
